Initial commit.
Working code for MPLAB v8.92 on a PIC32-MAXI-WEB for Ethernet TCP/IP
communication in the TCP DEMO PROJECT folder. Also includes a Microchip
folder with several libraries that may be unnecessary with this project.
I obtained the code from the Olimex website and have included it in the
DemoSoft_PIC32-MAXI-WEB_v1_03.zip file.