Typed keys are displayed on screen!
Typed keys are displayed on screen. When the system starts up it will be
in a test pattern. When a key is pressed, it will show that key on
screen (backward). The only keys working so far are q and w, every other
key produces a "!". The circular buffer for ps/2 keys works correctly,
but could use some cleaning up.