Added line input base and cursor limiting
Added a new function placeChar() to wrap writechar() and control where
things are on the screen. Needs work with the line in the cases of
having text before the line starts, and accumulating the line. A good
way to do accumulation is fill the line buffer, then copy it when you
hit enter.