Having issues with line wrap
trying to shift a character to the right is being harder than I thought
it would. Will have to figure out how to get exact offsets that I need
instead of guess and checking. Guess and check got me to be able to get
the character I need, just not in the right position.