I admit defeat. TCP/IP implementation is bad.
I can't seem to get a new socket to connect after disconnecting the
first one. I am not sure why this happens or how it happens, but it does
happen. Looking at information in wireshark simply tells me that the
second time does not operate like the first one. It appears that I send
the message from 240.0.0.0 instead of a local IP address 192.168.11.7
like it should. No idea what is causing this, I don't have a firm grip
on how TCP/IP works.